chore: combine thread page content component from temporary and chat conv
